move image_composition
