+++ packages/flame/test/image_composition_test.dart
@@ -1,4 +1,5 @@
-import 'package:flame/image_composition.dart';
+import 'package:flame/src/image_composition.dart';
+import 'package:flame_test/flame_test.dart';
 import 'package:mocktail/mocktail.dart';
 import 'package:test/test.dart';
 
@@ -6,32 +7,29 @@ class MockImage extends Mock implements Image {}
 
 void main() {
   group('ImageComposition', () {
-    group('add', () {
-      test('breaks assertion when adding an invalid portion', () {
-        final image = MockImage();
-        final composition = ImageComposition();
+    test('breaks assertion when adding an invalid portion', () {
+      final composition = ImageComposition();
+      final image = MockImage();
+      when(() => image.width).thenReturn(100);
+      when(() => image.height).thenReturn(100);
 
-        when(() => image.width).thenReturn(100);
-        when(() => image.height).thenReturn(100);
+      final invalidRects = [
+        const Rect.fromLTWH(-10, 10, 10, 10),
+        const Rect.fromLTWH(10, -10, 10, 10),
+        const Rect.fromLTWH(110, 10, 10, 10),
+        const Rect.fromLTWH(0, 110, 10, 10),
+        const Rect.fromLTWH(0, 0, 110, 110),
+        const Rect.fromLTWH(20, 0, 90, 10),
+        const Rect.fromLTWH(0, 20, 90, 90),
+        const Rect.fromLTWH(0, 0, 190, 90),
+        const Rect.fromLTWH(0, 0, 90, 190),
+      ];
 
-        final invalidRects = [
-          const Rect.fromLTWH(-10, 10, 10, 10),
-          const Rect.fromLTWH(10, -10, 10, 10),
-          const Rect.fromLTWH(110, 10, 10, 10),
-          const Rect.fromLTWH(0, 110, 10, 10),
-          const Rect.fromLTWH(0, 0, 110, 110),
-          const Rect.fromLTWH(20, 0, 90, 10),
-          const Rect.fromLTWH(0, 20, 90, 90),
-          const Rect.fromLTWH(0, 0, 190, 90),
-          const Rect.fromLTWH(0, 0, 90, 190),
-        ];
-
-        invalidRects.forEach((r) {
-          expect(
-            () => composition.add(image, Vector2.zero(), source: r),
-            throwsA(isA<AssertionError>()),
-          );
-        });
+      invalidRects.forEach((rect) {
+        expect(
+          () => composition.add(image, Vector2.zero(), source: rect),
+          failsAssert('Source rect should fit within the image'),
+        );
       });
     });
   });
